stop 64 bit conversion warnings

+++ stream.test.c
@@ -20,7 +20,7 @@ static void test_read(CuTest * tc, const stream * strm) {
     sz = strm->api->read(strm->handle, buf, sizeof(buf));
     CuAssertIntEquals(tc, 0, buf[sz]);
     CuAssertStrEquals(tc, "Hello\nWorld\n", buf);
-    CuAssertIntEquals(tc, 12, sz);
+    CuAssertIntEquals(tc, 12, (int)sz);
 }
 
 static void test_readln(CuTest * tc, const stream * strm) {
@@ -44,7 +44,7 @@ static void test_readln_unterminated(CuTest * tc, const stream * strm) {
     char buf[64];
     const char * hello = "Hello World";
 
-    CuAssertIntEquals(tc, 11, strm->api->write(strm->handle, hello, strlen(hello)));
+    CuAssertIntEquals(tc, 11, (int)strm->api->write(strm->handle, hello, strlen(hello)));
 
     strm->api->rewind(strm->handle);
     CuAssertIntEquals(tc, 0, strm->api->readln(strm->handle, buf, sizeof(buf)));
@@ -56,13 +56,13 @@ static void test_read_write(CuTest *tc, stream *strm)
     char ch;
 
     strm->api->rewind(strm->handle);
-    CuAssertIntEquals(tc, 3, strm->api->write(strm->handle, "123", 3));
+    CuAssertIntEquals(tc, 3, (int)strm->api->write(strm->handle, "123", 3));
     strm->api->rewind(strm->handle);
-    CuAssertIntEquals(tc, 1, strm->api->read(strm->handle, &ch, 1));
+    CuAssertIntEquals(tc, 1, (int)strm->api->read(strm->handle, &ch, 1));
     CuAssertIntEquals(tc, '1', ch);
-    CuAssertIntEquals(tc, 1, strm->api->read(strm->handle, &ch, 1));
+    CuAssertIntEquals(tc, 1, (int)strm->api->read(strm->handle, &ch, 1));
     CuAssertIntEquals(tc, '2', ch);
-    CuAssertIntEquals(tc, 1, strm->api->read(strm->handle, &ch, 1));
+    CuAssertIntEquals(tc, 1, (int)strm->api->read(strm->handle, &ch, 1));
     CuAssertIntEquals(tc, '3', ch);
 }
